+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- BlackBoxModbusServer base server task not being stopped before the derived object is destroyed.
+
 ## [2.1.1] - 2026-08-20
 ### Fixed
 - Out-of-bounds read in HardwareInterfaceConfigurationHR::OnWrite Wi-Fi password validation.